#142 Minor refactoring in watering logic
